Added Controller the controllers adding are to be continued

+++ Business/Concrete/ProfilePictureManager.cs
@@ -1,13 +1,15 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using Business.Abstruct;
 using Business.Constat;
 using Core.Utilities.BusinessRules;
+using Core.Utilities.Helpers.FileHelpers;
+using Core.Utilities.Helpers.FileHelpers.FileOnDiskManager;
 using Core.Utilities.Results.Abstruct;
 using Core.Utilities.Results.Concrute;
 using DataAccess.Abstruct;
 using Entities.Concrete;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore.Internal;
 
 namespace Business.Concrete
@@ -15,17 +17,22 @@ namespace Business.Concrete
     public class ProfilePictureManager : IProfilePictureService
     {
         private readonly IProfilePictureDal _profilePictureDal;
+        private readonly IUserService _userService;
+        private IFileSystem _fileSystem;
+        private readonly string _path = "ProfilePictures";
 
-        public ProfilePictureManager(IProfilePictureDal profilePictureDal)
+        public ProfilePictureManager(IProfilePictureDal profilePictureDal, IUserService userService, IFileSystem fileSystem)
         {
             _profilePictureDal = profilePictureDal;
+            _userService = userService;
+            _fileSystem = fileSystem;
         }
 
         public IDataResult<List<ProfilePicture>> GetAll()
         {
             IResult result = BusinessRule.Run
                 (
-                    CheckIfProfilePhotosExsist()
+                    CheckIfProfilePhotosExist()
                 );
             if (result != null)
             {
@@ -35,7 +42,7 @@ namespace Business.Concrete
             return new SuccessDataResult<List<ProfilePicture>>(_profilePictureDal.GetAll());
         }
 
-        private IResult CheckIfProfilePhotosExsist()
+        private IResult CheckIfProfilePhotosExist()
         {
             var result = _profilePictureDal.GetAll().Any();
             if (result)
@@ -46,7 +53,7 @@ namespace Business.Concrete
             return new ErrorResult(ProfilePhotoMessages.ThisProfilePhotosDoNotExist);
         }
 
-        public IDataResult<Entities.Concrete.ProfilePicture> GetById(int id)
+        public IDataResult<ProfilePicture> GetById(int id)
         {
             IResult result = BusinessRule.Run
                 (
@@ -60,6 +67,40 @@ namespace Business.Concrete
             return new SuccessDataResult<ProfilePicture>(_profilePictureDal.Get(pf => pf.Id == id));
         }
 
+        public IDataResult<List<ProfilePicture>> GetByUserId(int id)
+        {
+            IResult result = BusinessRule.Run
+                (
+                    _userService.CheckIfUserExist(id)
+                );
+
+            if (result != null)
+            {
+                return new ErrorDataResult<List<ProfilePicture>>(result.Message);
+            }
+
+            return new SuccessDataResult<List<ProfilePicture>>(_profilePictureDal.GetAll(pp=>pp.UserId == id));
+        }
+
+        public IResult Add(IFormFile file, ProfilePicture profilePicture)
+        {
+            IResult result = BusinessRule.Run
+                (
+                    CheckIfProfilePhotoAlreadyExist(profilePicture.Id),
+                    _userService.CheckIfUserExist(profilePicture.UserId)
+                );
+            if (result!= null)
+            {
+                return result;
+            }
+
+            profilePicture.PicturePath = _fileSystem.Add(file, _path);
+            profilePicture.Date = DateTime.Now;                                                                                                                                     
+            _profilePictureDal.Add(profilePicture);
+
+            return new SuccessResult();
+        }
+
         private IResult CheckIfProfilePhotoExist(int id)
         {
             var result = _profilePictureDal.GetAll(pf => pf.Id == id).Any();
@@ -69,20 +110,6 @@ namespace Business.Concrete
             }
 
             return new ErrorResult(ProfilePhotoMessages.ThisProfilePhotoDoNotExist);
-        }
-
-        public IResult Add(ProfilePicture profilePicture)
-        {
-            IResult result = BusinessRule.Run
-                (
-                    CheckIfProfilePhotoAlreadyExist(profilePicture.Id)
-                );
-            if (result != null)
-            {
-                return result;
-            }
-            _profilePictureDal.Add(profilePicture);
-            return new SuccessResult();
         }
 
         private IResult CheckIfProfilePhotoAlreadyExist(int profilePictureId)
@@ -102,11 +129,39 @@ namespace Business.Concrete
                 (
                 CheckIfProfilePhotoAlreadyDeleted(id)
                 );
+            if (result != null)
+            {
+                return result;
+            }
+
             var deleteToProfilePicture = _profilePictureDal.Get(pf => pf.Id == id);
+            new FileOnDiskManager().Delete(deleteToProfilePicture.PicturePath);
             _profilePictureDal.Delete(deleteToProfilePicture);
 
             return new SuccessResult();
 
+        }
+
+        public IResult Update(IFormFile file, ProfilePicture profilePicture)
+        {
+            IResult result = BusinessRule.Run
+                (
+                    CheckIfProfilePhotoExist(profilePicture.Id),
+                    _userService.CheckIfUserExist(profilePicture.UserId)
+                );
+            if (result != null)
+            {
+                return result;
+            }
+
+            var profilePictureToUpdate = _profilePictureDal.Get(pp => pp.Id == profilePicture.Id);
+
+            profilePicture.Id = profilePictureToUpdate.Id;
+            profilePicture.PicturePath =
+                new FileOnDiskManager().Update(file, profilePictureToUpdate.PicturePath, _path);
+            profilePicture.Date = DateTime.Now;
+
+            return new SuccessResult();
         }
 
         private IResult CheckIfProfilePhotoAlreadyDeleted(int id)
@@ -121,20 +176,5 @@ namespace Business.Concrete
             return new ErrorResult(ProfilePhotoMessages.ThisProfilePhotoAlreadyDeleted);
         }
 
-
-        public IResult Update(ProfilePicture profilePicture)
-        {
-            IResult result = BusinessRule.Run
-                (
-                    CheckIfProfilePhotoExist(profilePicture.Id)
-                );
-            if (result != null)
-            {
-                return result;
-            }
-
-            _profilePictureDal.Update(profilePicture);
-            return new SuccessResult();
-        }
     }
 }
